+++ src/field_camera.c
@@ -99,8 +99,8 @@ void DrawWholeMapView(void)
 
 static void DrawWholeMapViewInternal(int x, int y, const struct MapLayout *mapLayout)
 {
-    u8 i;
-    u8 j;
+    u32 i;
+    u32 j;
     u32 r6;
     u8 temp;
 
@@ -137,7 +137,7 @@ static void RedrawMapSlicesForCameraUpdate(struct FieldCameraOffset *cameraOffse
 
 static void RedrawMapSliceNorth(struct FieldCameraOffset *cameraOffset, const struct MapLayout *mapLayout)
 {
-    u8 i;
+    u32 i;
     u8 temp;
     u32 r7;
 
@@ -156,7 +156,7 @@ static void RedrawMapSliceNorth(struct FieldCameraOffset *cameraOffset, const st
 
 static void RedrawMapSliceSouth(struct FieldCameraOffset *cameraOffset, const struct MapLayout *mapLayout)
 {
-    u8 i;
+    u32 i;
     u8 temp;
     u32 r7 = cameraOffset->yTileOffset * 32;
 
@@ -171,7 +171,7 @@ static void RedrawMapSliceSouth(struct FieldCameraOffset *cameraOffset, const st
 
 static void RedrawMapSliceEast(struct FieldCameraOffset *cameraOffset, const struct MapLayout *mapLayout)
 {
-    u8 i;
+    u32 i;
     u8 temp;
     u32 r6 = cameraOffset->xTileOffset;
 
@@ -186,7 +186,7 @@ static void RedrawMapSliceEast(struct FieldCameraOffset *cameraOffset, const str
 
 static void RedrawMapSliceWest(struct FieldCameraOffset *cameraOffset, const struct MapLayout *mapLayout)
 {
-    u8 i;
+    u32 i;
     u8 temp;
     u8 r5 = cameraOffset->xTileOffset + 28;
